[go_router] Fix return type of current state getter to be non-nullable (#8173)
The `state` getter returns a `GoRouterState?`, but I don't think it needs to be nullable since `currentConfiguration.last.buildState()` never returns `null`.

I'm not sure if it is nullable on purpose for some changes planned in the Future. If there is a reason, it might be better to add a comment about it. Otherwise, it is just inconvenient that a null check is necessary.