[video_player_avfoundation] iOS platform view support (#8237)
This PR adds support for platform views on iOS as a way of displaying a
video. When creating a video, it's now possible to choose between
texture view approach (rendered using `Texture` widget on the Flutter
side) and platform view approach (rendered on the native side, using
`AVPlayerLayer`).

`FVPVideoPlayer` class now has nothing to do with texture. The
texture-related code was moved from it to `FVPTextureBasedVideoPlayer` -
a subclass of `FVPVideoPlayer` that adds texture functionality. In the
plugin class (`createWithOptions` method) we create either the basic
version (for platform view) or the texture subclass (in case of texture
approach) based on the parameter passed in from Flutter side.

Platform view is only supported on iOS, no MacOS implementation is added
in this PR. The functionality is not yet exposed in the app-facing
package (only in example app) - it will be done later, once we add the
Android implementation. The PR does not introduce breaking changes, I
followed the rule "non-breaking changes, even at the expense of a
less-clean API" (`buildViewWithOptions` and `createWithOptions`
methods).

Up to this point, the variable naming relied heavily on the texture (we
had a lot of `textureId` variables and properties). Since now you can
use a platform view instead of a texture view, these variables and
parameters were renamed to just `playerId`.

I left some comments in the PR to clarify/discuss some choices.

Resolves [#86613](https://github.com/flutter/flutter/issues/86613) (not
fully though, as it is not yet available in the app-facing package).

## Pre-launch Checklist

- [x] I read the [Contributor Guide] and followed the process outlined
there for submitting PRs.
- [x] I read the [Tree Hygiene] page, which explains my
responsibilities.
- [x] I read and followed the [relevant style guides] and ran the
auto-formatter. (Unlike the flutter/flutter repo, the flutter/packages
repo does use `dart format`.)
- [x] I signed the [CLA].
- [x] The title of the PR starts with the name of the package surrounded
by square brackets, e.g. `[shared_preferences]`
- [x] I [linked to at least one issue that this PR fixes] in the
description above.
- [x] I updated `pubspec.yaml` with an appropriate new version according
to the [pub versioning philosophy], or this PR is [exempt from version
changes].
- [x] I updated `CHANGELOG.md` to add a description of the change,
[following repository CHANGELOG style], or this PR is [exempt from
CHANGELOG changes].
- [x] I updated/added relevant documentation (doc comments with `///`).
- [x] I added new tests to check the change I am making, or this PR is
[test-exempt].
- [x] All existing and new tests are passing.

If you need help, consider asking for advice on the #hackers-new channel
on [Discord].

<!-- Links -->
[Contributor Guide]:
https://github.com/flutter/packages/blob/main/CONTRIBUTING.md
[Tree Hygiene]:
https://github.com/flutter/flutter/blob/master/docs/contributing/Tree-hygiene.md
[relevant style guides]:
https://github.com/flutter/packages/blob/main/CONTRIBUTING.md#style
[CLA]: https://cla.developers.google.com/
[Discord]:
https://github.com/flutter/flutter/blob/master/docs/contributing/Chat.md
[linked to at least one issue that this PR fixes]:
https://github.com/flutter/flutter/blob/master/docs/contributing/Tree-hygiene.md#overview
[pub versioning philosophy]: https://dart.dev/tools/pub/versioning
[exempt from version changes]:
https://github.com/flutter/flutter/blob/master/docs/ecosystem/contributing/README.md#version
[following repository CHANGELOG style]:
https://github.com/flutter/flutter/blob/master/docs/ecosystem/contributing/README.md#changelog-style
[exempt from CHANGELOG changes]:
https://github.com/flutter/flutter/blob/master/docs/ecosystem/contributing/README.md#changelog
[test-exempt]:
https://github.com/flutter/flutter/blob/master/docs/contributing/Tree-hygiene.md#tests