[various] Remove plugin-level `integration_test` dependencies (#8711)
Many of our plugins have `dev_dependencies` on `integration_test` at the plugin level, for some reason. Integration tests run in the example app, not the plugin, so only the example app should need the dependency.

In practice it shouldn't matter since it's a dev dependency, but it's best to be accurate about our requirements.