[camera_android] Don't override default fps range when not recording (#8891)
In production I have an app using the `camera_android` plugin that just uses the frame streaming feature. I've been pinning version `0.10.8+18` for a while because we found an issue after updating to `0.10.9`. Some users reported a dark camera preview. 

After some investigation, I believe the issue is because of the call to `this.cameraFeatures.setFpsRange` in the `Camera` constructor. That is the only difference in behavior that .9 and .8+18 have.

Fixes  https://github.com/flutter/flutter/issues/165491

From the linked issue I went with the second possible fix. Would that be a good approach? 
Or should the fps setting also apply to the camera streaming and preview?
cc @camsim99 

## Pre-Review Checklist

[^1]: Regular contributors who have demonstrated familiarity with the repository guidelines only need to comment if the PR is not auto-exempted by repo tooling.